requirements.txt
python-telegram-bot>=20.0
ccxt>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
flask>=2.0.0
